refactor: move plugin/theme data functions to separate files
