+++ src/types.ts
@@ -16,3 +16,23 @@ export interface VersionHistory {
 	version: string;
 	initialReleaseDate: string;
 }
+
+export interface DownloadDataPoint {
+	date: string;
+	downloads?: number;
+	growth?: number;
+}
+
+export interface PerMonthDataPoint {
+	year: string;
+	month: string;
+	value: number;
+}
+
+export interface DownloadReleaseCorrelationDataPoint {
+	id: string;
+	downloads: number;
+	releases: number;
+	initialReleaseDate: number;
+	initialReleaseDateString: string;
+}
