+++ website/src/components/plugins/pluginsCount.astro
@@ -1,7 +1,7 @@
 ---
 import PluginDataFile from '../../../../plugin-data.json' assert { type: 'json' };
-import { getPluginCountPerMonth, getPluginCountPerMonthWoRetiredPlugins, getRetiredPluginCountPerMonth } from '../../utils/utils';
 import PluginPerMonthChart from '../svelte/dataPerMonthChart.svelte';
+import { getPluginCountPerMonth, getPluginCountPerMonthWoRetiredPlugins, getRetiredPluginCountPerMonth } from "../../../../src/plugin/data";
 
 const dataPoints = getPluginCountPerMonth(PluginDataFile);
 const dataPoints2 = getPluginCountPerMonthWoRetiredPlugins(PluginDataFile);
