+++ website/src/components/plugins/pluginsDownloadCorrelations.astro
@@ -1,5 +1,5 @@
 ---
-import { getDownloadReleaseCorrelationDataPoints } from '../../utils/utils';
+import { getDownloadReleaseCorrelationDataPoints } from "../../../../src/plugin/data";
 import PluginDataFile from '../../../../plugin-data.json' assert { type: 'json' };
 import PluginReleaseDownloadsCorrelationChart from '../svelte/pluginReleaseDownloadsCorrelationChart.svelte';
