+++ website/src/components/plugins/pluginsDownloadStats.astro
@@ -1,5 +1,5 @@
 ---
-import { getDownloadDataPoints } from '../../utils/utils';
+import { getDownloadDataPoints } from "../../../../src/plugin/data";
 import PluginDataFile from '../../../../plugin-data.json' assert { type: 'json' };
 import PluginDownloadChart from '../svelte/pluginDownloadChart.svelte';
