+++ website/src/components/plugins/pluginsNewReleasesPerMonth.astro
@@ -1,6 +1,6 @@
 ---
 import PluginDataFile from '../../../../plugin-data.json' assert { type: 'json' };
-import { getNewPluginReleasesPerMonth } from '../../utils/utils';
+import { getNewPluginReleasesPerMonth } from "../../../../src/plugin/data";
 import DataPerMonthChart from '../svelte/dataPerMonthChart.svelte';
 
 const dataPoints = getNewPluginReleasesPerMonth(PluginDataFile);
