+++ website/src/components/plugins/pluginsRetiredAll.astro
@@ -1,6 +1,6 @@
 ---
 import PluginDataFile from '../../../../plugin-data.json' assert { type: 'json' };
-import { getRetiredPlugins } from '../../utils/utils';
+import { getRetiredPlugins } from "../../../../src/plugin/data";
 import PluginLink from '../pluginLink.astro';
 import Commit from '../commit.astro';
