+++ website/src/components/plugins/pluginsRetiredPercentageByReleaseMonth.astro
@@ -1,6 +1,6 @@
 ---
 import PluginDataFile from '../../../../plugin-data.json' assert { type: 'json' };
-import { getPercentageOfRetiredPluginsByReleaseMonth } from '../../utils/utils';
+import { getPercentageOfRetiredPluginsByReleaseMonth } from "../../../../src/plugin/data";
 import DataPerMonthChart from '../svelte/dataPerMonthChart.svelte';
 
 const dataPoints = getPercentageOfRetiredPluginsByReleaseMonth(PluginDataFile);
