+++ website/src/components/plugins/pluginsRetiredRecent.astro
@@ -1,6 +1,6 @@
 ---
 import PluginDataFile from '../../../../plugin-data.json' assert { type: 'json' };
-import { getRecentRetiredPlugins } from '../../utils/utils';
+import { getRecentRetiredPlugins } from "../../../../src/plugin/data";
 import PluginLink from '../pluginLink.astro';
 import Commit from '../commit.astro';
