+++ website/src/components/themes/themesNewReleasesPerMonth.astro
@@ -1,6 +1,6 @@
 ---
 import ThemeDataFile from '../../../../theme-data.json' assert { type: 'json' };
-import { getNewThemeReleasesPerMonth } from '../../utils/utils';
+import { getNewThemeReleasesPerMonth } from "../../../../src/theme/data";
 import DataPerMonthChart from '../svelte/dataPerMonthChart.svelte';
 
 const dataPoints = getNewThemeReleasesPerMonth(ThemeDataFile);
