+++ website/src/components/themes/themesRetiredAll.astro
@@ -1,6 +1,6 @@
 ---
 import ThemeDataFile from '../../../../theme-data.json' assert { type: 'json' };
-import { getRetiredThemes } from '../../utils/utils';
+import { getRetiredThemes } from "../../../../src/theme/data";
 import Commit from '../commit.astro';
 import ThemeLink from '../themeLink.astro';
