+++ website/src/components/themes/themesRetiredPerMonth.astro
@@ -1,6 +1,6 @@
 ---
 import ThemeDataFile from '../../../../theme-data.json' assert { type: 'json' };
-import { getRetiredThemesPerMonth } from '../../utils/utils';
+import { getRetiredThemesPerMonth } from "../../../../src/theme/data";
 import DataPerMonthChart from '../svelte/dataPerMonthChart.svelte';
 
 const dataPoints = getRetiredThemesPerMonth(ThemeDataFile);
