+++ website/src/components/themes/themesRetiredPercentageByReleaseMonth.astro
@@ -1,6 +1,6 @@
 ---
 import ThemeDataFile from '../../../../theme-data.json' assert { type: 'json' };
-import { getPercentageOfRetiredThemesByReleaseMonth } from '../../utils/utils';
+import { getPercentageOfRetiredThemesByReleaseMonth } from "../../../../src/theme/data";
 import DataPerMonthChart from '../svelte/dataPerMonthChart.svelte';
 
 const dataPoints = getPercentageOfRetiredThemesByReleaseMonth(ThemeDataFile);
