+++ website/src/components/themes/themesRetiredRecent.astro
@@ -1,6 +1,6 @@
 ---
 import ThemeDataFile from '../../../../theme-data.json' assert { type: 'json' };
-import { getRecentRetiredThemes } from '../../utils/utils';
+import { getRecentRetiredThemes } from "../../../../src/theme/data";
 import Commit from '../commit.astro';
 import ThemeLink from '../themeLink.astro';
